+++ src/parser/inventory/weapon.js
@@ -203,13 +203,15 @@ let getDamage = (data, flags, betterRolls5e) => {
   const twoHanded = data.definition.properties.find((property) => property.name === "Two-Handed");
   const mod = (offHand && !twoWeapon) ? "" : " + @mod";
   const damageType = getDamageType(data);
+  const globalDamageHints = game.settings.get("ddb-importer", "use-damage-hints");
+  const damageHint = damageType && globalDamageHints ? `[${damageType}]` : "";
 
   const versatile = data.definition.properties
     .filter((property) => property.name === "Versatile")
     .map((versatile) => {
       if (versatile && versatile.notes) {
         return (
-          utils.parseDiceString(versatile.notes + ` + ${magicalDamageBonus}`, null, `[${damageType}]`, greatWeaponFighting).diceString + mod
+          utils.parseDiceString(versatile.notes + ` + ${magicalDamageBonus}`, null, damageHint, greatWeaponFighting).diceString + mod
         );
       } else {
         return "";
@@ -239,7 +241,7 @@ let getDamage = (data, flags, betterRolls5e) => {
 
     // if there is a magical damage bonus, it probably should only be included into the first damage part.
     parts.push([
-      utils.parseDiceString(diceString + ` + ${magicalDamageBonus}`, `${mod}${dueling}`, `[${damageType}]`, fightingStyleDiceMod)
+      utils.parseDiceString(diceString + ` + ${magicalDamageBonus}`, `${mod}${dueling}`, damageHint, fightingStyleDiceMod)
         .diceString,
         damageType,
     ]);
@@ -251,7 +253,8 @@ let getDamage = (data, flags, betterRolls5e) => {
     .forEach((mod) => {
       const damagePart = (mod.dice) ? mod.dice.diceString : mod.value;
       if (damagePart) {
-        const damageParsed = utils.parseDiceString(damagePart, "", `[${mod.subType}]`).diceString;
+        const damageHintSub = mod.subType && globalDamageHints ? `[${mod.subType}]` : "";
+        const damageParsed = utils.parseDiceString(damagePart, "", damageHintSub).diceString;
         parts.push([damageParsed, mod.subType]);
       }
     });
@@ -263,7 +266,7 @@ let getDamage = (data, flags, betterRolls5e) => {
     .forEach((mod) => {
       const damagePart = (mod.dice) ? mod.dice.diceString : `${mod.value}`;
       if (damagePart) {
-        const subType = mod.subType ? `[${mod.subType}]` : "";
+        const subType = mod.subType && globalDamageHints ? `[${mod.subType}]` : "";
         const damageParsed = utils.parseDiceString(damagePart, "", subType).diceString;
         if (utils.isModuleInstalledAndActive("betterrolls5e")) {
           const attackNum = parts.length;